+++ run_python_tests.sh
@@ -124,7 +124,7 @@ fi
 
 declare -A FLOOR_DEVELOPER=(
     ["examples/ardep_ecu"]=312
-    ["examples/basic_ecu"]=426
+    ["examples/basic_ecu"]=573
     ["examples/basic_ecu_doip"]=99
     ["examples/basic_ecu_doip_freertos"]=99
     ["examples/basic_ecu_freertos"]=99
@@ -137,7 +137,7 @@ declare -A FLOOR_DEVELOPER=(
 )
 declare -A FLOOR_PROFESSIONAL=(
     ["examples/ardep_ecu"]=450
-    ["examples/basic_ecu"]=482
+    ["examples/basic_ecu"]=629
     ["examples/basic_ecu_doip"]=155
     ["examples/basic_ecu_doip_freertos"]=155
     ["examples/basic_ecu_freertos"]=155